scripts/requirements.txt
numpy==1.14.3
Pillow==5.1.0
